% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_handling.R
\name{getLib}
\alias{getLib}
\title{Returns lib in form of a list.}
\usage{
getLib(name)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Returns lib in form of a list.
}
\examples{
add(1, 1)
add(10, 1)
}
