% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{queryChea3}
\alias{queryChea3}
\title{Query ChEA3}
\usage{
queryChea3(set_name = "untitled", geneset, n_results = 10,
  libnames = names(libs), integrate_libs = T, method = c("FET"),
  background = NULL)
}
\arguments{
\item{set_name}{Name of query. Defaults to "untitled".}

\item{geneset}{Character array of gene set to query}

\item{n_results}{Number of results to return for each TF regulon library. Defaults to 10.}

\item{libnames}{Character array of TF regulon library names to use. Defaults to all libraries available in the package.}

\item{integrate_libs}{Integrate library results. Often results in a better transcription factor ranking. Defaults to TRUE.}

\item{method}{Method(s) for comparison of TF regulons and query. Defaults to Fisher's Exact Test}

\item{background}{The size of the universe of genes under consideration. User may set to a constant. Default is NULL which results in background being set to the size of the union of the unique genes of the query and TF regulon library.}
}
\value{
A list of dataframes with the top TF regulon hits for each TF regulon library tested.
}
\description{
Compute enrichment of a gene set query over TF regulon libraries
}
\examples{
data(example_query)
results = queryChea3(set_name = "exp1_DEG",example_query)

}
